# src/main/resources/static/swagger-ui/openapi3.yaml
openapi: 3.0.1
info:
  title: 801base API docs
  description: 801base의 API 문서입니다.
  version: 0.0.1
servers:
- url: https://api2.801base.com
...
paths:
...
  /admin/products:
    get:
      tags:
      - Admin-Product
      summary: 상품 목록을 조회할 수 있다.
      operationId: 관리자_상품_목록_조회_성공
      parameters:
      - name: name
        in: query
        description: 상품명
        required: false
        schema:
          type: string
      - name: intensityId
        in: query
        description: 원두 강도 카테고리 ID
        required: false
        schema:
          type: string
      - name: cupSizeId
        in: query
        description: 컵 사이즈 카테고리 ID
        required: false
        schema:
          type: string
      - name: status
        in: query
        description: "판매상태(ON_SALE, UNAVAILABLE, HIDDEN)"
        required: false
        schema:
          type: string
      - name: page
        in: query
        description: "페이지 번호 (1부터 시작, 기본값: 1)"
        required: false
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-products-1101392596"
              examples:
                관리자_상품_목록_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "content" : [ {
                          "id" : 1,
                          "name" : "콜드브루",
                          "price" : 3500,
                          "quantity" : 100,
                          "thumbnail" : "https://test.com/thumbnail.png",
                          "intensity" : "Strong",
                          "cupSize" : "Large",
                          "status" : "ON_SALE"
                        } ],
                        "page" : 1,
                        "size" : 10,
                        "totalPages" : 1,
                        "totalElements" : 1
                      },
                      "error" : null
                    }
    post:
      tags:
      - Admin-Product
      summary: 상품을 등록할 수 있다.
      description: |2-
                        PRD-101: 상품명이 비어있습니다.
                        PRD-102: 상품명이 너무 깁니다.
                        PRD-103: 가격은 양수여야합니다.
                        PRD-104: 재고는 음수일 수 없습니다.
                        PRD-105: 유효하지 않은 카테고리입니다.
      operationId: 상품_등록_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/admin-products-247955186"
            examples:
              상품_등록_성공:
                value: |-
                  {
                    "name" : "콜드브루",
                    "price" : 3500,
                    "quantity" : 100,
                    "thumbnail" : "https://test.com/thumbnail.png",
                    "detailImage" : "https://test.com/detailImage.png",
                    "intensityId" : 1,
                    "cupSizeId" : 10
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-products9820101"
              examples:
                상품_등록_성공:
                  value: |-
                    {
                      "data" : {
                        "productId" : 10
                      },
                      "error" : null
                    }
...
  /admin/products/selling-status:
    get:
      tags:
      - Admin-Product
      summary: 상품 판매상태 목록을 조회할 수 있다.
      operationId: 상품_판매상태_조회_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-products-selling-status-152159338"
              examples:
                상품_판매상태_조회_성공:
                  value: |-
                    {
                      "data" : [ {
                        "code" : "ON_SALE",
                        "label" : "판매중"
                      } ],
                      "error" : null
                    }
  /admin/products/{productId}:
    get:
      tags:
      - Admin-Product
      summary: 상품 상세 정보를 조회할 수 있다.
      description: |2-
                        PRO-001: 상품을 찾을 수 없습니다.
                        PRO-002: 상품 재고를 찾을 수 없습니다.
      operationId: 관리자_상품_상세_조회_성공
      parameters:
      - name: productId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-products-productId-412506376"
              examples:
                관리자_상품_상세_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "id" : 1,
                        "name" : "콜드브루",
                        "price" : 3500,
                        "quantity" : 100,
                        "thumbnail" : "https://test.com/thumbnail.png",
                        "detailImage" : "https://test.com/detail.png",
                        "intensity" : "Strong",
                        "cupSize" : "Large",
                        "status" : "ON_SALE"
                      },
                      "error" : null
                    }
    put:
      tags:
      - Admin-Product
      summary: 상품을 수정할 수 있다.
      description: |2-
                        PRO-001: 상품을 찾을 수 없습니다.
                        PRO-002: 상품 재고를 찾을 수 없습니다.
                        PRD-101: 상품명이 비어있습니다.
                        PRD-102: 상품명이 너무 깁니다.
                        PRD-103: 가격은 양수여야합니다.
                        PRD-104: 재고는 음수일 수 없습니다.
                        PRD-105: 유효하지 않은 카테고리입니다.
      operationId: 상품_수정_성공
      parameters:
      - name: productId
        in: path
        description: 상품 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/admin-products-productId-1789436874"
            examples:
              상품_수정_성공:
                value: |-
                  {
                    "name" : "콜드브루",
                    "price" : 3500,
                    "quantity" : 100,
                    "thumbnail" : "https://test.com/thumbnail.png",
                    "detailImage" : "https://test.com/detailImage.png",
                    "intensityId" : 1,
                    "cupSizeId" : 10,
                    "status" : "UNAVAILABLE"
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-products-productId-1712578160"
              examples:
                상품_수정_성공:
                  value: |-
                    {
                      "data" : {
                        "productId" : 10
                      },
                      "error" : null
                    }
    delete:
      tags:
      - Admin-Product
      summary: 상품을 삭제할 수 있다.
      description: |2-
                        PRO-001: 상품을 찾을 수 없습니다.
                        PRO-002: 상품 재고를 찾을 수 없습니다.
      operationId: 상품_삭제_성공
      parameters:
      - name: productId
        in: path
        description: 상품 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                상품_삭제_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "OK"
                      },
                      "error" : null
                    }
...
  /cart-items:
    get:
      tags:
      - Cart
      summary: 장바구니에 담긴 상품 목록을 조회할 수 있다.
      operationId: 장바구니_상품_목록_조회_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cart-items-761185737"
              examples:
                장바구니_상품_목록_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "totalPrice" : 80000,
                        "deliveryPrice" : 0,
                        "cartItems" : [ {
                          "cartItemId" : 101,
                          "productId" : 1,
                          "productName" : "Product 1",
                          "quantity" : 2,
                          "price" : 10000,
                          "stockQuantity" : 10,
                          "thumbnail" : "thumbnail1.jpg",
                          "isAvailable" : true
                        }, {
                          "cartItemId" : 102,
                          "productId" : 2,
                          "productName" : "Product 2",
                          "quantity" : 3,
                          "price" : 20000,
                          "stockQuantity" : 5,
                          "thumbnail" : "thumbnail2.jpg",
                          "isAvailable" : true
                        } ]
                      },
                      "error" : null
                    }
    post:
      tags:
      - Cart
      summary: 사용자의 장바구니에 상품을 추가할 수 있다.
      operationId: 장바구니에_상품_추가_성공
      parameters:
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/cart-items1349440338"
            examples:
              장바구니에_상품_추가_성공:
                value: |-
                  {
                    "productId" : 1,
                    "quantity" : 100
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cart-items151658045"
              examples:
                장바구니에_상품_추가_성공:
                  value: |-
                    {
                      "data" : {
                        "quantity" : 80,
                        "stockQuantity" : 80,
                        "requiresQuantityAdjustment" : true
                      },
                      "error" : null
                    }
    delete:
      tags:
      - Cart
      summary: 장바구니에 추가된 상품을 삭제할 수 있다.
      operationId: 장바구니_상품_삭제_성공
      parameters:
      - name: cartItemIds
        in: query
        description: 상품 아이디
        required: false
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/admin-reviews-reviewId-reply1686665306"
              examples:
                장바구니_상품_삭제_성공:
                  value: |-
                    {
                      "data" : {
                        "message" : "Successfully deleted 1 cart items"
                      },
                      "error" : null
                    }
  /cart-items/{cartItemId}:
    patch:
      tags:
      - Cart
      summary: 장바구니 내 상품의 수량을 변경할 수 있다.
      operationId: 장바구니_상품_수정_성공
      parameters:
      - name: cartItemId
        in: path
        description: 장바구니 상품 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: Authorization
        required: true
        schema:
          type: string
        example: Bearer sample-token
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/cart-items-cartItemId-2049702389"
            examples:
              장바구니_상품_수정_성공:
                value: |-
                  {
                    "cartItemId" : 1,
                    "quantity" : 100
                  }
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/cart-items-cartItemId838557620"
              examples:
                장바구니_상품_수정_성공:
                  value: |-
                    {
                      "data" : {
                        "userId" : 1,
                        "productId" : 1,
                        "quantity" : 80,
                        "stockQuantity" : 80,
                        "requiresQuantityAdjustment" : true
                      },
                      "error" : null
                    }
...
  /products/{productId}:
    get:
      tags:
      - Product
      summary: 상품 상세 정보를 조회할 수 있다.
      description: |2-
                        PRO-001: 상품을 찾을 수 없습니다.
                        PRO-002: 상품 재고를 찾을 수 없습니다.
      operationId: 상품_상세_조회_성공
      parameters:
      - name: productId
        in: path
        description: 상품 ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/products-productId-988628468"
              examples:
                상품_상세_조회_성공:
                  value: |-
                    {
                      "data" : {
                        "id" : 1,
                        "name" : "콜드브루",
                        "price" : 3500,
                        "quantity" : 100,
                        "thumbnail" : "https://test.com/thumbnail.png",
                        "detailImage" : "https://test.com/detail.png",
                        "intensity" : "Strong",
                        "cupSize" : "Large"
                      },
                      "error" : null
                    }
...
components:
  schemas:
...
    admin-products-productId-412506376:
      type: object
      properties:
        data:
          required:
          - cupSize
          - detailImage
          - id
          - intensity
          - name
          - price
          - quantity
          - status
          - thumbnail
          type: object
          properties:
            intensity:
              type: string
              description: 원두 강도
            thumbnail:
              type: string
              description: 썸네일 이미지 URL
            quantity:
              type: number
              description: 재고 수량
            price:
              type: number
              description: 가격
            name:
              type: string
              description: 상품명
            id:
              type: number
              description: 상품 ID
            cupSize:
              type: string
              description: 컵 사이즈
            status:
              type: string
              description: 판매상태
            detailImage:
              type: string
              description: 상세 이미지 URL
    cart-items-761185737:
      type: object
      properties:
        data:
          required:
          - cartItems
          - deliveryPrice
          - totalPrice
          type: object
          properties:
            totalPrice:
              type: number
              description: 총 상품 금액 (배송비 제외)
            deliveryPrice:
              type: number
              description: 배송비
            cartItems:
              type: array
              description: 장바구니 상품 목록
              items:
                required:
                - cartItemId
                - isAvailable
                - price
                - productId
                - productName
                - quantity
                - stockQuantity
                - thumbnail
                type: object
                properties:
                  isAvailable:
                    type: boolean
                    description: 구매 가능 여부
                  thumbnail:
                    type: string
                    description: 상품 썸네일 이미지 URL
                  quantity:
                    type: number
                    description: 수량
                  productId:
                    type: number
                    description: 상품 ID
                  price:
                    type: number
                    description: 상품 단가
                  stockQuantity:
                    type: number
                    description: 재고 수량
                  productName:
                    type: string
                    description: 상품명
                  cartItemId:
                    type: number
                    description: 장바구니 아이템 ID
...
    products-productId-988628468:
      type: object
      properties:
        data:
          required:
          - cupSize
          - detailImage
          - id
          - intensity
          - name
          - price
          - quantity
          - thumbnail
          type: object
          properties:
            intensity:
              type: string
              description: 원두 강도
            thumbnail:
              type: string
              description: 썸네일 이미지 URL
            quantity:
              type: number
              description: 재고 수량
            price:
              type: number
              description: 가격
            name:
              type: string
              description: 상품명
            id:
              type: number
              description: 상품 ID
            cupSize:
              type: string
              description: 컵 사이즈
            detailImage:
              type: string
              description: 상세 이미지 URL
...
    admin-products-247955186:
      required:
      - cupSizeId
      - detailImage
      - intensityId
      - name
      - price
      - quantity
      - thumbnail
      type: object
      properties:
        thumbnail:
          type: string
          description: 썸네일 이미지
        intensityId:
          type: number
          description: 강도 카테고리 아이디
        quantity:
          type: number
          description: 재고 수량
        price:
          type: number
          description: 가격
        cupSizeId:
          type: number
          description: 컵사이즈 카테고리 아이디
        name:
          type: string
          description: 상품명
        detailImage:
          type: string
          description: 상세 이미지
    cart-items151658045:
      type: object
      properties:
        data:
          required:
          - quantity
          - requiresQuantityAdjustment
          - stockQuantity
          type: object
          properties:
            quantity:
              type: number
              description: 상품 수량
            requiresQuantityAdjustment:
              type: boolean
              description: 상품 수량이 재고 수량을 초과함
            stockQuantity:
              type: number
              description: 재고 수량
...
    cart-items1349440338:
      required:
      - productId
      - quantity
      type: object
      properties:
        quantity:
          type: number
          description: 상품수량
        productId:
          type: number
          description: 상품Id
...
    cart-items-cartItemId838557620:
      type: object
      properties:
        data:
          required:
          - productId
          - quantity
          - requiresQuantityAdjustment
          - stockQuantity
          - userId
          type: object
          properties:
            quantity:
              type: number
              description: 수량
            productId:
              type: number
              description: 상품 아이디
            requiresQuantityAdjustment:
              type: boolean
              description: 수량 변경 여부
            stockQuantity:
              type: number
              description: 재고 수량
            userId:
              type: number
              description: 유저 아이디
...
    admin-products9820101:
      type: object
      properties:
        data:
          required:
          - productId
          type: object
          properties:
            productId:
              type: number
              description: 생성된 상품 아이디
...
    admin-reviews-reviewId-reply1686665306:
      type: object
      properties:
        data:
          required:
          - message
          type: object
          properties:
            message:
              type: string
              description: 응답 메시지
...
    cart-items-cartItemId-2049702389:
      required:
      - cartItemId
      - quantity
      type: object
      properties:
        quantity:
          type: number
          description: 수량
        cartItemId:
          type: number
          description: 카트 아이디
...
    admin-products-productId-1789436874:
      required:
      - cupSizeId
      - detailImage
      - intensityId
      - name
      - price
      - quantity
      - status
      - thumbnail
      type: object
      properties:
        thumbnail:
          type: string
          description: 썸네일 이미지
        intensityId:
          type: number
          description: 강도 카테고리 아이디
        quantity:
          type: number
          description: 재고 수량
        price:
          type: number
          description: 가격
        cupSizeId:
          type: number
          description: 컵사이즈 카테고리 아이디
        name:
          type: string
          description: 상품명
        status:
          type: string
          description: 판매상태 코드
        detailImage:
          type: string
          description: 상세 이미지
...
    admin-products-1101392596:
      type: object
      properties:
        data:
          required:
          - page
          - size
          - totalElements
          - totalPages
          type: object
          properties:
            size:
              type: number
              description: 페이지 크기
            totalPages:
              type: number
              description: 전체 페이지 수
            page:
              type: number
              description: 현재 페이지 번호
            content:
              type: array
              items:
                required:
                - cupSize
                - id
                - intensity
                - name
                - price
                - quantity
                - status
                - thumbnail
                type: object
                properties:
                  intensity:
                    type: string
                    description: 원두 강도
                  thumbnail:
                    type: string
                    description: 썸네일 이미지 URL
                  quantity:
                    type: number
                    description: 재고
                  price:
                    type: number
                    description: 가격
                  name:
                    type: string
                    description: 상품명
                  id:
                    type: number
                    description: 상품 ID
                  cupSize:
                    type: string
                    description: 컵 사이즈
                  status:
                    type: string
                    description: 판매상태
            totalElements:
              type: number
              description: 총 상품 수
...
    admin-products-selling-status-152159338:
      type: object
      properties:
        data:
          type: array
          items:
            required:
            - code
            - label
            type: object
            properties:
              code:
                type: string
                description: 판매상태 코드
              label:
                type: string
                description: 판매상태
...
    admin-products-productId-1712578160:
      type: object
      properties:
        data:
          required:
          - productId
          type: object
          properties:
            productId:
              type: number
              description: 수정된 상품 아이디
...